Conflicts:
	compiler/cli/src/org/jetbrains/kotlin/cli/jvm/compiler/JetCoreEnvironment.java

+++ j2k/testData/fileOrElement/literalExpression/endsWithDFL.kt
@@ -1,21 +1,21 @@
 class Test {
     fun test() {
-        val l1 = 10
+        val l1 = 10L
         val d1 = 10.0
-        val f1 = 10.0.toFloat()
+        val f1 = 10.0f
 
-        val l2 = 10
+        val l2 = 10L
         val d2 = 10.0
-        val f2 = 10.0.toFloat()
+        val f2 = 10.0f
     }
 
     fun testBoxed() {
-        val l1 = 10
+        val l1 = 10L
         val d1 = 10.0
-        val f1 = 10.0.toFloat()
+        val f1 = 10.0f
 
-        val l2 = 10
+        val l2 = 10L
         val d2 = 10.0
-        val f2 = 10.0.toFloat()
+        val f2 = 10.0f
     }
 }
